make health status provider registration more dynamic
Now all healthStatus providers are checked on bundle events, if they are
stil valid (== their corresponding mbean is still present).
Also on bundle events the config is checked, if new config can be
activated.

If a bundle introduces new mbeans, for which a config already exists,
the healthstatus providers are created.