add JCRObservationDelayProvider
The JCRObservationDelay provider makes it possible to monitor the delay
between triggering an observaiton in the JCR (e.g. saving a node) and
the handling of the respective event through JCR observation.

This provider writes periodically a node in the JCR repository and
listens for the observation for this action. The time difference is
tracked and can be categorized into WARN and CRITICAL by configuring
arbitrary thresholds via OSGI.